Refactoring, moving base dir and splitting validation-engine
