--- src/main/resources/etc/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "raml-shapes",
-  "version": "0.0.1",
-  "description": "RAML Shapes Validator",
-  "main": "shapes.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "rdflib": "^0.15.0"
-  }
-}
