+++ src/test/resources/sh/tests/amf/basic2.jsonld
@@ -1,0 +1,1 @@
+{"http://raml.org/vocabularies/shapes#toValidate":{"http://raml.org/vocabularies/shapes/anon#name":{"@value":"John"}}}
