+++ src/test/resources/sh/tests/amf/basicShapes2.jsonld
@@ -1,0 +1,76 @@
+{
+  "@type": [
+    "http://www.w3.org/ns/shacl#NodeShape",
+    "http://www.w3.org/ns/shacl#Shape"
+  ],
+  "@id": "http://raml.org/vocabularies/data#AnonShape",
+  "http://www.w3.org/ns/shacl#property": [
+    {
+      "http://www.w3.org/ns/shacl#maxCount": [
+        {
+          "@value": 1
+        }
+      ],
+      "http://www.w3.org/ns/shacl#dataType": [
+        {
+          "@id": "http://www.w3.org/2001/XMLSchema#string"
+        }
+      ],
+      "@id": "http://raml.org/vocabularies/data#AnonShape/property/name",
+      "@type": [
+        "http://www.w3.org/ns/shacl#PropertyShape",
+        "http://www.w3.org/ns/shacl#Shape"
+      ],
+      "http://www.w3.org/ns/shacl#path": [
+        {
+          "@id": "http://raml.org/vocabularies/shapes/anon#name"
+        }
+      ],
+      "http://raml.org/vocabularies/shapes#propertyLabel": [
+        {
+          "@value": "name"
+        }
+      ],
+      "http://www.w3.org/ns/shacl#minCount": [
+        {
+          "@value": 1
+        }
+      ]
+    },
+    {
+      "http://www.w3.org/ns/shacl#maxCount": [
+        {
+          "@value": 1
+        }
+      ],
+      "http://www.w3.org/ns/shacl#dataType": [
+        {
+          "@id": "http://www.w3.org/2001/XMLSchema#string"
+        }
+      ],
+      "@id": "http://raml.org/vocabularies/data#AnonShape/property/surname",
+      "@type": [
+        "http://www.w3.org/ns/shacl#PropertyShape",
+        "http://www.w3.org/ns/shacl#Shape"
+      ],
+      "http://www.w3.org/ns/shacl#path": [
+        {
+          "@id": "http://raml.org/vocabularies/shapes/anon#surname"
+        }
+      ],
+      "http://raml.org/vocabularies/shapes#propertyLabel": [
+        {
+          "@value": "surname"
+        }
+      ],
+      "http://www.w3.org/ns/shacl#minCount": [
+        {
+          "@value": 1
+        }
+      ]
+    }
+  ],
+  "http://www.w3.org/ns/shacl#targetObjectsOf": {
+    "@id": "http://raml.org/vocabularies/shapes#toValidate"
+  }
+}
